Web UI: Rename Logs→Monitor, extract Token Estimation section, improve Raw YAML editor (#18)
* Initial plan

* Fix Web UI issues: rename Logs to Monitor, add Token Estimation section, improve Raw YAML editor


---------
