+++ web/src/components/config/ConfigSidebar.tsx
@@ -10,6 +10,7 @@ export type ConfigSectionId =
   | 'models'
   | 'channels'
   | 'prompt'
+  | 'token'
   | 'routing'
   | 'raw';
 
@@ -66,6 +67,11 @@ const sectionIcons: Record<ConfigSectionId, React.ReactNode> = {
       <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={1.5} d="M8 10h.01M12 10h.01M16 10h.01M9 16H5a2 2 0 01-2-2V6a2 2 0 012-2h14a2 2 0 012 2v8a2 2 0 01-2 2h-5l-5 5v-5z" />
     </svg>
   ),
+  token: (
+    <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+      <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={1.5} d="M9 7h6m0 10v-3m-3 3h.01M9 17h.01M9 14h.01M12 14h.01M15 11h.01M12 11h.01M9 11h.01M7 21h10a2 2 0 002-2V5a2 2 0 00-2-2H7a2 2 0 00-2 2v14a2 2 0 002 2z" />
+    </svg>
+  ),
   routing: (
     <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
       <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={1.5} d="M13 7h8m0 0v8m0-8l-8 8-4-4-6 6" />
@@ -80,7 +86,7 @@ const sectionIcons: Record<ConfigSectionId, React.ReactNode> = {
 
 const sectionIds: ConfigSectionId[] = [
   'global', 'datasource', 'concurrency', 'columns', 'validation',
-  'models', 'channels', 'prompt', 'routing',
+  'models', 'channels', 'prompt', 'token', 'routing',
 ];
 
 export default function ConfigSidebar({ activeSection, onSelect, language }: ConfigSidebarProps) {
@@ -95,6 +101,7 @@ export default function ConfigSidebar({ activeSection, onSelect, language }: Con
     models: t.cfgModels,
     channels: t.cfgChannels,
     prompt: t.cfgPrompt,
+    token: t.cfgTokenEstimation,
     routing: t.cfgRouting,
     raw: t.rawYamlEditor,
   };
